[MAJOR UPDATE] Geode LX driver
This is a major re-vamping of the Geode LX driver to support EXA, RandR,
and advanced xserver features.  The GX and LX drivers now share the similar
streamlined infrastructure, and acceleration.  The LX driver also adds
features the GX does not have, such as accelerated rotations and
far better composite support.

This provides the basis for the rotation and acceleration enhancements.