Several fixes to support OLPC LX turnon
These are fixes discovered during the OLPC LX turn on.  The compression
was incorrectly being configured, resulting in bad drawing, that has
bee repaired.  Assume by default that we're going to use a panel and the
CRT, no longer read the straps MSR because it is unreliable.  Finally,
fix the MSR hooks by moving the hook definitions where they will do some
good and disabling the VSA method.