Fix the upload / composite mask race
We fix a race with the upload function when using a composite mask
by using the exa core function exaGetPixmapFirstPixel, which does
the right thing in X 1.4 and newer.  For older versions, the fix is
to stall the pipeine to ensure the upload is complete, so thats what
we do.