[GX][LX] Push PreInit fixes to avoid segfauls on X -configure.
This is known to work on products with GeodeROM and Insyde BIOSes.

On products with a General Software BIOS, a freeze occurs while X
is probing VBE.  The same issue occurs with LinuxBIOS and VGAROM,
but not with LinuxBIOS omitting VGAROM.

Both issues appear to be caused by X server core upgrading from
vm86 to x86emu since X server release 1.3, which requires fixing
x86emu, or the concerned BIOSes, or both.