Fix build for git-master (missing device pointers).
We don't have an updated Video ABI yet, so use the XINPUT ABI as check for
when to use which call.

Note that the check is not optimal. We only check for the position of the VCP,
not of any other device to adjust the viewpoint. This needs to be fixed.