xf86-video-geode: change cursor size to match cimarron
When we use ARGB cursors and rotate the screen, we get double cursors
and stale cursor data left over.  This is because cimarron expects a
cursor that is smaller than or equal to 48x64, but hardcodes a memory
shift of 192 bytes (4*48).  This means that if we have a 32x64 cursor,
cimarron screws it up.  Change the cursor size to 48x64.

Signed-off-by: Andres Salomon <dilinger@debian.org>