xf86-video-geode: fix up cursor size (again)
The geode hardware (and cimarron) expects cursor memory size to be 48x64,
and cimarron will zero out the excess when a smaller cursor is used.  It
would be nice to be able to use the full cursor size, but xorg's interleaving
stuff requires a cursor width that's a multiple of 32.  Thus, we're stuck
back at 32x32 for cursor size, but we allocate enough memory for a 48x64
cursor.

This fixes misc interleaving corruption that we see with 2bpp cursors (when
converting to ARGB).

Signed-off-by: Andres Salomon <dilinger@debian.org>