Change OLPC detection heuristic to one that works on upstream kernels
We've been checking whether we have a DCON/are on an OLPC by looking for
files from the "olpc_dcon" kernel module.  This module isn't upstream yet,
so this only works for the OLPC custom kernel.  We might as well check for
an OLPC kernel module that *is* upstream, so that we can run on unmodified
distributions.  This patch moves us to checking for "olpc-ac" instead.

Signed-off-by: Chris Ball <cjb@laptop.org>