Fix GPIO BAR detection with libpciaccess
This patch fixes an issue with the GPIO BAR detection.  Basically,
with libpciaccess we're finding the ISA device and checking its BARs,
but we're not bothering to fill in the BAR information via
pci_device_probe.  This results in a 0'd out bar and:

(EE) GEODE(0): Could not find the GPIO I/O base

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>
Signed-off-by: Andres Salomon <dilinger@collabora.co.uk>
Signed-off-by: Chris Ball <cjb@laptop.org>