lx_exa: fallback in case of server-side gradients and solid fill
RENDER extension version 0.10 added support for doing server-side
gradients and solid fills, which we were not handling at all.
cairo-1.9 started making use of server-side gradients, and we are
crashing on those, as the source pixmap (pxSrc in lx_prepare_composite)
doesn't have a drawable in this case, and we were not expecting such
a possibility.
So, as the first measure, fallback (three years late) all server-side
gradients and solid fills instead of crashing, until we have no code
that tries to accelerate cases of these or aren't sure if we even
can accelerate any cases.

pSourcePict member was added to xserver in 2005, so no compatibility
wrapping needed.

Signed-off-by: Mart Raudsepp <leio@gentoo.org>
Acked-by: Otavio Salvador <otavio@ossystems.com.br>