Put the one pixel rendering work back to the server to handle if the pMsk is not zero
Put the one pixel rendering work back to the server to handle if the pMsk is not zero

*exaScratch.srcWidth and exaScratch.srcHeight are used to reocrd
mask's width and mask's height if mask is not zero. So the one pixel
source's width and height are missing in lx_do_composite. So we must
fallback. Otherwise, bigger region will be wrongly rendered.

Signed-off-by: Frank Huang <frankr.huang@amd.com>