fix the DCON verification loop for LX output
This is pretty clearly a bug. This should fix it (after all, that
check is merely to see if the panel is a DCON; we don't care at
all about the panel bit). This also adds an extra parenthesis in
the following if() statement for clarity.

I'm resisting the temptation to change GeodeRec's Output member to
an unsigned long (for now). Bitfields should really be unsigned.

Signed-off-by: Andres Salomon <dilinger@queued.net>