Improve handling of gamma correction in video vs graphics
The LX is a bit odd in that the palette can be used for gamma correction
of graphics data, or video data, but not both at the same time.

Right now, during X startup, X sets gamma correction (but without any
actual correction) causing the Geode driver to start using the palette
for graphics gamma correction. (this is just momentary)

Later on during startup, vg_set_custom_mode() switches to using the
palette for video gamma correction, and this is the end result.

If you later use xrandr to change the gamma, the palette starts being used
for graphics again.

If you VT switch, vg_set_custom_mode (momentarily) starts using the palette
for video, but then X jumps in with a gamma callback causing it to be used
for graphics again.

End result: no user visible bug, but this is inconsistent.
As there is no exposed mechanism for changing video gamma I suggest
we just drop that bit from vg_set_custom_mode(). Then the inconsistency
goes away.

Downside: this now becomes dependent on X setting gamma during startup.
I don't know when this was introduced but I think it is recent.

Aditionally, I think I found a (harmless) bug in
df_set_video_palette_entry, it doesn't do what the comments say.