Fix packed overlay offscreen allocations
Commit 5e72a00ad2 caused packed video data to corrupt glyphs and other
parts of the screen, but it turns out that the commit actually at fault
was d681a844e, incorrectly changing the size of the allocated destination
memory from the number of bytes needed to the number of lines needed.

While fixing this, I noticed that LXAllocateSurface is probably making the
same mistake, and that the height is probably not calculated correctly for
some corner cases when calling LXCopyFromSys in the packed
video path (but I'm not sure about either).

Fixes https://bugs.freedesktop.org/show_bug.cgi?id=33004