Remove unrequired large file descriptor partial support
The LFS "transitional extension" API is deprecated and is not available on FreeBSD.
Large file support (64 bit) is not required on the geode 32 bit only architecture.

There are some hints that the LFS transional extension API were not used
correclty. The variable holding the value is held in an unsigned long rather
than in off_t. The msr open call did not use the O_LARGEFILE flag and did not
check for EOVERFLOW.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>