Move ztv driver to it's own directory
The convention is to have the video driver (geode) in the src subdir
and any other targets (utils, apps, other drivers) in their own
directory. It makes it much simpler to right a correct makefile
and to set conditional compile.

This highlighted a few things to fix. For now, the directory includes
have not been copied as they are required. The AM_CCASFLAGS is not required
either as there is no assembler.

Other than that, the gcc command issued has no differences. Other unrequired
defines could be removed later.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>