Add a configure option to enable/disable building the ztv driver
Even when V4L2 is available there are always reasons why distros
may not want to build the driver.

This patch uses a common idiom in xorg.
When no configure option is specified ztv is built iff v4l2 is detected (auto).
When user issues --disable-ztv, the ztv driver is not built (no)
When user issues --enable-ztv, the ztv driver is built (yes) if v4l2 is detected
but the configuration fails if v4l2 is missing. Distros do not want silent
failures when a feature is explicitly requested but cannot be built.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>