README: do not suggest compiling with xserver older than 1.3
This was done at one point in time by some drivers (3 that I remember)
but the workarounds all have been removed since. The tarballs created
by those drivers picked-up whatever code happened to be on the
developer disk and was not in git. The code was not maintained
and in fact did not work thereafter.

http://cgit.freedesktop.org/xorg/driver/xf86-video-intel/commit/configure.ac
?id=84f69081abaeef8b05cafb64d3102eb2abdf9a8e

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>