Geode 2.11.13
This release mainly features a complete overhaul of the building scripts plus
a cleanup of deprecated macros. The key benefit is that this Geode driver can
finally build on FreeBSD and on other platforms not offering V4L2 support and
that it can also build on a 64-bit host using its 32-bit personality.

Support for the Video Input Port (VIP) feature of the Geode LX found in 'ztv'
is now documented and the source code saw the removal of deprecated Linux 2.4
backward compatibility code. Compiling is skipped on platforms without V4L2.

Configuration for laptops featuring WXGA resolutions is finally documented.

Setting the GEODE_TRACE_FALL macro now enables composite operation tracing on
this driver. This can be used to profile the performance during development.

Fixes to keep this driver compilable on recent X servers are also included.

VALIDATION PLATFORMS
* Debian/stable (X server 1.7.7) on Artec DBE61 and Hercules EC-800.
* Ubuntu/Precise (X server 1.10.4) on FIC ION603A.

Signed-off-by: Martin-Éric Racine <martin-eric.racine@iki.fi>