lx_exa: Fix crash with solid fills on PictOpAdd operations
lx_check_composite has many special checks for PictOpAdd to return
successfully early, but these were done even earlier than the check
for solid fills. This resulted in lx_prepare_composite assuming
the source pixmap exists, which is not the case with solid fills.
Move the solid fill fallback checks before the others, so we are
guaranteed to always have a source pixmap in PrepareComposite and
Composite and don't crash trying to access a NULL pointer.

Also move up the mask check to do less checks in the PictOpAdd logic.