lx_exa: Implement solid pictures support as source with a mask
cairo-1.12 uses solid pictures instead of 1x1R pixmaps in glyph rendering
paths, so accelerate it.
In addition to acceleration, it avoids a bug in xserver-1.13.0 and earlier
which causes visible misrendering for fallback path, making cairo-1.12 a
viable and desired choice on GeodeLX systems.
Quick benchmarking suggests a 4-12% win in cairo-traces.