Add config.h include to src/lx_memory.c
The src/lx_memory.c file uses the xf86.h header file.  This file must have
HAVE_STRNDUP defined before calling it when the building platform has a strndup
function.  When using config.h, this file doesn't have that define and so fails
to compile.

The attached patch adds the conditional config.h to this file so it may compile
on the affected platforms.

The patch is trivial and may be included and used under whatever licensing
desired.

Closes: Free Desktop Bug #84541

Signed-off-by: Brian A. Lloyd <brian.lloyd@familyhonor.net>