LX: Restore old VT switch code
Calls for enabling and disabling outputs were removed from the VT switch
routine when RandR support was added. Presumably, this is because the
server already sets DPMS properly on a VT switch. However, DDX code can
undo what the server does so it seems best to keep the explicit calls.

Closes: X.Org Bugzilla #80825

Signed-off-by: Connor Behan <connor.behan@gmail.com>