Suppress majority of compiler warnings
This applies some obvious changes to stop gcc from complaining about
dead code, shadow declarations, differing signedness and sections that
mix declarations with code.

The warning about discarding const qualifiers is more annoying because
we pass string literals to some functions which accept non-const
pointers. Currently, this takes the following approach. If the function
*needs* to accept non-const pointers, cast the string literal as char *.
Otherwise, change the function to only accept a const pointer. To
anticipate future use cases though, I could also leave function
definitions as they are and just always cast string literals.
Alternatively, if this is more trouble that it is worth, we could just
put up with the warnings.

Signed-off-by: Connor Behan <connor.behan@gmail.com>